Periodically Check and Close Game Launch Threads (#2859)
* fix isNumeric for strings longer than max int

* periodically check and close launch threads

+++ src/main/java/ti4/helpers/ButtonHelper.java
@@ -5245,7 +5245,7 @@ public class ButtonHelper {
         if (strNum == null)
             return false;
         try {
-            Integer.parseInt(strNum);
+            Long.parseLong(strNum);
             return true;
         } catch (NumberFormatException nfe) {
             return false;
